Fix glitch with decimal point
An error previously occurred with the decimal point where it would
display even if there's already another decimal point on the screen. to
fix it, instead of holding a true/false function in a variable, the
program checks if there is already a period in the output, and makes a
decision based off of that.